+++ examples/derivatives_trading_portfolio_margin/rest_api/trade_api/new_um_algo_order.rs
@@ -1,5 +1,4 @@
 use anyhow::{Context, Result};
-use rust_decimal::prelude::*;
 use std::env;
 use tracing::info;
 
@@ -37,7 +36,6 @@ async fn main() -> Result<()> {
         "BNBUSDT".to_string(),
         NewUmAlgoOrderSideEnum::Buy,
         NewUmAlgoOrderTypeEnum::Stop,
-        dec!(0.01),
     )
     .build()?;