+++ src/algo/rest_api/apis/future_algo_api.rs
@@ -292,15 +292,21 @@ impl std::str::FromStr for VolumeParticipationFutureAlgoPositionSideEnum {
 ///
 /// This struct holds all of the inputs you can pass when calling
 /// [`cancel_algo_order_future_algo`](#method.cancel_algo_order_future_algo).
-#[derive(Clone, Debug, Builder, Deserialize)]
+#[derive(Clone, Debug, Builder, Deserialize, Default)]
 #[builder(pattern = "owned", build_fn(error = "ParamBuildError"))]
 pub struct CancelAlgoOrderFutureAlgoParams {
     /// eg. 14511
     ///
-    /// This field is **required.
-    #[builder(setter(into))]
-    #[serde(rename = "algoId")]
-    pub algo_id: i64,
+    /// This field is **optional.
+    #[builder(setter(into), default)]
+    #[serde(rename = "algoId", default)]
+    pub algo_id: Option<i64>,
+    /// eg. "65ce1630101a480b85915d7e11fd5078"
+    ///
+    /// This field is **optional.
+    #[builder(setter(into), default)]
+    #[serde(rename = "clientAlgoId", default)]
+    pub client_algo_id: Option<String>,
     /// Request validity window in milliseconds
     ///
     /// This field is **optional.
@@ -312,13 +318,9 @@ pub struct CancelAlgoOrderFutureAlgoParams {
 impl CancelAlgoOrderFutureAlgoParams {
     /// Create a builder for [`cancel_algo_order_future_algo`].
     ///
-    /// Required parameters:
-    ///
-    /// * `algo_id` — eg. 14511
-    ///
     #[must_use]
-    pub fn builder(algo_id: i64) -> CancelAlgoOrderFutureAlgoParamsBuilder {
-        CancelAlgoOrderFutureAlgoParamsBuilder::default().algo_id(algo_id)
+    pub fn builder() -> CancelAlgoOrderFutureAlgoParamsBuilder {
+        CancelAlgoOrderFutureAlgoParamsBuilder::default()
     }
 }
 /// Request parameters for the [`query_current_algo_open_orders_future_algo`] operation.
@@ -637,13 +639,20 @@ impl FutureAlgoApi for FutureAlgoApiClient {
     ) -> anyhow::Result<RestApiResponse<models::CancelAlgoOrderFutureAlgoResponse>> {
         let CancelAlgoOrderFutureAlgoParams {
             algo_id,
+            client_algo_id,
             recv_window,
         } = params;
 
         let mut query_params = BTreeMap::new();
         let body_params = BTreeMap::new();
 
-        query_params.insert("algoId".to_string(), json!(algo_id));
+        if let Some(rw) = algo_id {
+            query_params.insert("algoId".to_string(), json!(rw));
+        }
+
+        if let Some(rw) = client_algo_id {
+            query_params.insert("clientAlgoId".to_string(), json!(rw));
+        }
 
         if let Some(rw) = recv_window {
             query_params.insert("recvWindow".to_string(), json!(rw));
@@ -1134,7 +1143,7 @@ mod tests {
         TOKIO_SHARED_RT.block_on(async {
             let client = MockFutureAlgoApiClient { force_error: false };
 
-            let params = CancelAlgoOrderFutureAlgoParams::builder(1).build().unwrap();
+            let params = CancelAlgoOrderFutureAlgoParams::builder().build().unwrap();
 
             let resp_json: Value =
                 serde_json::from_str(r#"{"algoId":14511,"success":true,"code":0,"msg":"OK"}"#)
@@ -1158,7 +1167,9 @@ mod tests {
         TOKIO_SHARED_RT.block_on(async {
             let client = MockFutureAlgoApiClient { force_error: false };
 
-            let params = CancelAlgoOrderFutureAlgoParams::builder(1)
+            let params = CancelAlgoOrderFutureAlgoParams::builder()
+                .algo_id(1)
+                .client_algo_id("65ce1630101a480b85915d7e11fd5078".to_string())
                 .recv_window(5000)
                 .build()
                 .unwrap();
@@ -1185,7 +1196,7 @@ mod tests {
         TOKIO_SHARED_RT.block_on(async {
             let client = MockFutureAlgoApiClient { force_error: true };
 
-            let params = CancelAlgoOrderFutureAlgoParams::builder(1).build().unwrap();
+            let params = CancelAlgoOrderFutureAlgoParams::builder().build().unwrap();
 
             match client.cancel_algo_order_future_algo(params).await {
                 Ok(_) => panic!("Expected an error"),